Port almost done. (~50%)
Started porting my repo todo-list-app to SvelteKit to allow my app to be more scalable.

Activities management is working like a charm after a bit of thinking.
Added code to disable the drawer if the local database isn't started.

It's working so far and it's kinda cool to be honest :)